add allure report
